+++ lib/src/screen/home_screen.dart
@@ -9,7 +9,6 @@ import 'package:kakao_map_plugin_example/src/screen/login.dart';
 import 'package:kakao_map_plugin_example/src/screen/my_page.dart';
 import 'package:kakao_map_plugin_example/src/screen/review_list.dart';
 import 'package:kakao_map_plugin_example/src/screen/review_write.dart';
-import 'package:kakao_map_plugin_example/src/screen/test_mypage.dart';
 import 'package:kakao_map_plugin_example/src/service/geolocator_service.dart';
 import 'package:kakao_map_plugin_example/src/widget/app_bar.dart';
 import 'package:kakao_map_plugin_example/src/widget/home_button.dart';
@@ -54,7 +53,7 @@ class _HomeScreenState extends State<HomeScreen>
         preferredSize: const Size.fromHeight(50),
         child: CustomAppBar(
           title: _selectedIndex == 0
-              ? 'Hansajang'
+              ? '한사장'
               : _selectedIndex == 1
                   ? "Search"
                   : _selectedIndex == 2
@@ -64,11 +63,11 @@ class _HomeScreenState extends State<HomeScreen>
       ),
       //body 바꿀거임 ㅇㅇ
       body: _selectedIndex == 0
-          ? homeContainer()
+          ? const IndexScreen()
           : _selectedIndex == 1
               ? testContainer()
               : _selectedIndex == 2
-                  ? const TestMyPage()
+                  ? const MyPage()
                   : testHome(),
       bottomNavigationBar: SizedBox(
         height: 90,
@@ -107,76 +106,6 @@ class _HomeScreenState extends State<HomeScreen>
     return Container(
       child: const Center(child: Text("hi")),
     );
-  }
-
-  Widget homeContainer() {
-    return Center(
-      child: Stack(
-        children: [
-          Center(
-            child: Ink(
-              width: 300,
-              height: 480,
-              decoration: BoxDecoration(
-                boxShadow: [
-                  BoxShadow(
-                    color: Colors.grey.withOpacity(0.7),
-                    spreadRadius: 5,
-                    blurRadius: 5.0,
-                    offset: const Offset(0, 0),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          Center(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    testButton("1"),
-                    testButton("2"),
-                  ],
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    testButton("3"),
-                    testButton("4"),
-                  ],
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    testButton("5"),
-                    testButton("6"),
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget testButton(text) {
-    return Center(
-        child: Ink(
-      color: Colors.white,
-      child: InkWell(
-        onTap: () {
-          print(text);
-        },
-        child: Ink(
-          width: 150,
-          height: 160,
-          child: Text(text),
-        ),
-      ),
-    ));
   }
 
   // 기존에 있던 홈
